add policies de modulos existentes, add recursos de roles, configuracion de roles y permisos finalizado, modificacion de seeders para llenar datos de usuarios y roles, solucion de acoplamiento de profile y rol

+++ database/migrations/2025_03_02_165355_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name',100);
             $table->string('barcode',100);
             $table->text('description');
-            $table->integer('cost');
+            $table->decimal('cost', 10, 2);
             $table->decimal('price', 10, 2);
             $table->integer('stock');
             $table->integer('alerts');
